feat: network specific quick actions
