+++ src/client/src/views/home/account/AccountInfo.tsx
@@ -19,6 +19,7 @@ import { FedimintGatewayCard } from './gateway/FedimintGatewayCard';
 import { useGatewayEcashTotal } from '../../../hooks/UseGatewayEcashTotal';
 import { useGatewayState } from '../../../context/GatewayContext';
 import { GatewayInfo } from '../../../api/types';
+import { getNetworkIndicator } from './network';
 
 const S = {
   grid: styled.div<{ gatewayInfo?: GatewayInfo | null }>`
@@ -83,7 +84,7 @@ export const AccountInfo = () => {
     <>
       <CardWithTitle>
         <SubTitle>
-          Network: {gatewayInfo?.network ? gatewayInfo.network : 'Bitcoin'}{' '}
+          Network: {getNetworkIndicator(gatewayInfo?.network || 'bitcoin')}
         </SubTitle>
         <Card>
           <ResponsiveLine>
